+++ python/cudf/cudf/core/column/column.py
@@ -75,7 +75,6 @@ from cudf.utils.utils import (
     _array_ufunc,
     _is_null_host_scalar,
     is_na_like,
-    mask_dtype,
 )
 
 if TYPE_CHECKING:
@@ -647,7 +646,7 @@ class ColumnBase(Serializable, BinaryOperand, Reducible):
                 raise ValueError(f"Unsupported mode: {mode}")
         else:
             obj = None
-        return cuda.as_cuda_array(obj).view(mask_dtype)
+        return cuda.as_cuda_array(obj).view(SIZE_TYPE_DTYPE)
 
     def __len__(self) -> int:
         return self.size
