+++ python/cudf/cudf/core/mixins/__init__.py
@@ -1,7 +1,15 @@
-# Copyright (c) 2022, NVIDIA CORPORATION.
+# Copyright (c) 2022-2025, NVIDIA CORPORATION.
 
 from .binops import BinaryOperand
+from .getitem import GetAttrGetItemMixin
+from .notiterable import NotIterable
 from .reductions import Reducible
 from .scans import Scannable
 
-__all__ = ["BinaryOperand", "Reducible", "Scannable"]
+__all__ = [
+    "BinaryOperand",
+    "GetAttrGetItemMixin",
+    "NotIterable",
+    "Reducible",
+    "Scannable",
+]
